% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{conscious_route}
\alias{conscious_route}
\title{conscious_route}
\usage{
conscious_route(net, start = NULL, stop = NULL, gender_weight = 10)
}
\arguments{
\item{net}{Street network with gender column, returned by
\link{gender_streetnet} function.}

\item{start}{Start point of route.}

\item{stop}{End point of route.}

\item{gender_weight}{A numeric factor >= 1 by which to preferentially route
along streets named after women. Larger values will search larger distances
relative to the default (shortest) route to find routes with greatest
proportional distances along streets named after women. Arbitrarily large
values may be specified.}
}
\description{
Route between nominated start and end points along greatest proportional
length of streets named after women rather than men.
}
