% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse-city.R
\name{gcr_city}
\alias{gcr_city}
\title{gcr_city}
\usage{
gcr_city(net, wt_profile = "foot", n = NULL, quiet = FALSE)
}
\arguments{
\item{net}{Street network in \code{osmdata_sc} format.}

\item{wt_profile}{Type of weighting profile for street network routing,.}

\item{n}{Number of pints used to calculate pair-wise routes along which to
aggregate statistics. If no value specified, routes are calculated between
all pairs for the entire network. If value exceeds total number of network
nodes, that number is used instead.}

\item{quiet}{if \code{FALSE}, display progress information on screen.}
}
\description{
Analyse entire properties of one city
}
