% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender-streetnet.R
\name{gender_streetnet}
\alias{gender_streetnet}
\title{gender_streetnet}
\usage{
gender_streetnet(net, wt_profile = "foot", country = "de")
}
\arguments{
\item{net}{Street network in \pkg{sf} or \pkg{sc} format}

\item{wt_profile}{Type of weighting to be used; see \code{dodgr::weight_streetnet}
for details.}

\item{country}{Country for which gender is to be determined, either as
ISO3166 two-letter abbreviation, or full text. See \link{list_countries} for
list of recognized countries.}
}
\value{
A weighted street network in \pkg{dodgr} format.
}
\description{
Weight a street network via \pkg{dodgr} function \code{weight_streetnet}, adding
additional information on gender of street names
}
