% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendercorrectrouting.R
\docType{package}
\name{gendercorrectrouting}
\alias{gendercorrectrouting}
\alias{gendercorrectrouting-package}
\title{gendercorrectrouting}
\description{
Gender-Correct Routing Along Streets Named After Women Rather Than Men
}
