% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender.R
\name{get_gender}
\alias{get_gender}
\title{get_gender}
\usage{
get_gender(text, country)
}
\arguments{
\item{text}{The text string}

\item{country}{The country code for the text}
}
\value{
Gender of the text
}
\description{
Determine the gender of a given text string
}
