% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gender.R
\name{list_countries}
\alias{list_countries}
\title{list_countries}
\usage{
list_countries()
}
\value{
A \code{data.frame} object of country names and associated ISO3166
abbreviations.
}
\description{
List all countries included in the first name gender matcher
}
