% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/router.R
\name{conscious_route}
\alias{conscious_route}
\title{conscious_route}
\usage{
conscious_route(net, start, stop)
}
\arguments{
\item{net}{Street network with gender column, returned by
\link{gender_streetnet} function.}

\item{start}{Start point of route.}

\item{stop}{End point of route.}
}
\description{
Route between nominated start and end points along greatest proportional
length of streets named after women rather than men.
}
