% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderconsciousrouting.R
\docType{package}
\name{genderconsciousrouting}
\alias{genderconsciousrouting}
\alias{genderconsciousrouting-package}
\title{genderconsciousrouting}
\description{
Gender-Correct Routing Along Streets Named After Women Rather Than Men
}
