% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-gender.R
\name{get_gender}
\alias{get_gender}
\title{get_gender}
\usage{
get_gender(text)
}
\arguments{
\item{text}{The character vector to encode}
}
\value{
Vector of name length with gender codes
}
\description{
Get the gender of a character vector
}
